Function stun_server::stun_handler::error_response

fn error_response(
    error_code: usize, 
    reason: String, 
    transaction_id: [u8; 12]
) -> Vec<u8>

Creates bytes for a error message binding request

Creates error headers, attribute headers and error message, then serializes it using bincode and serde with big endian configuration.

Examples

let transaction_id = [ec, cb, 9a, 1a, aa, a9, cd, 9, 0, 0, 0, 0]
let bytes = error_response(500, String::from("Foo bar"), transaction_id);
println!("{:#x?}", bytes);
// Expected output: [1, 11, 0, 10, 21, 12, a4, 42, ec, cb, 9a, 1a, aa, a9, cd, 9, 0, 0, 0, 0, 0, 9, 0, a, 0, 0, 0, 1, f4, 46 6f 6f 20 62 61 72]